requirements.txt
Flask==2.3.2
google-cloud-datastore==2.15.0
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1
protobuf==3.20.*
six==1.16.0
flask-cors
python-jose
